+++ twine/commands/__init__.py
@@ -3,6 +3,7 @@
 The contents of this package are not a public API. For more details, see
 https://github.com/pypa/twine/issues/194 and https://github.com/pypa/twine/issues/665.
 """
+
 # Copyright 2013 Donald Stufft
 #
 # Licensed under the Apache License, Version 2.0 (the "License");