+++ twine/exceptions.py
@@ -1,4 +1,5 @@
 """Module containing exceptions raised by twine."""
+
 # Copyright 2015 Ian Stapleton Cordasco
 #
 # Licensed under the Apache License, Version 2.0 (the "License");