+++ twine/wheel.py
@@ -16,6 +16,7 @@ import os
 import re
 import zipfile
 from typing import List, Optional
+from typing import cast as type_cast
 
 from pkginfo import distribution
 
@@ -72,20 +73,27 @@ class Wheel(distribution.Distribution):
                 "Not a known archive format for file: %s" % fqn
             )
 
+        searched_files: List[str] = []
         try:
             for path in self.find_candidate_metadata_files(names):
                 candidate = "/".join(path)
                 data = read_file(candidate)
                 if b"Metadata-Version" in data:
                     return data
+                searched_files.append(candidate)
         finally:
             archive.close()
 
-        raise exceptions.InvalidDistribution("No METADATA in archive: %s" % fqn)
+        raise exceptions.InvalidDistribution(
+            "No METADATA in archive or METADATA missing 'Metadata-Version': "
+            "%s (searched %s)" % (fqn, ",".join(searched_files))
+        )
 
     def parse(self, data: bytes) -> None:
         super().parse(data)
 
         fp = io.StringIO(data.decode("utf-8", errors="replace"))
+        # msg is ``email.message.Message`` which is a legacy API documented
+        # here: https://docs.python.org/3/library/email.compat32-message.html
         msg = distribution.parse(fp)
-        self.description = msg.get_payload()
+        self.description = type_cast(str, msg.get_payload())